Fix flex error on the main and heroes wrapper
