+++ src/containers/ComicList.js
@@ -5,6 +5,8 @@ import { fetchComics, filterByCreator } from '../actions';
 import ComicCard from '../components/ComicCard';
 import Loading from '../components/Loading';
 import Error from '../components/Error';
+import { listOfCreators } from '../common';
+import FilterComics from '../components/FilterComics';
 
 const CharacterList = ({
   filter, comics, fetchComics, filterByCreator,
@@ -14,19 +16,28 @@ const CharacterList = ({
     fetchComics();
     filterByCreator('All');
   }, [fetchComics]);
+  const handleOnChnageSelect = e => {
+    filterByCreator(e.target.value);
+  };
 
   const displayComics = () => comics.map(comic => {
     const arrayOfUsers = comic.creators.items.map(name => name.name);
     if (arrayOfUsers.includes(filter) || filter === 'All') {
       return (
-        <ComicCard
-          title={comic.title}
-          thumbnail={comic.thumbnail}
-          format={comic.format}
-          key={comic.id}
-          creators={comic.creators}
-          id={comic.id}
-        />
+        <div>
+          <FilterComics
+            cretorList={listOfCreators}
+            onChangeSelect={handleOnChnageSelect}
+          />
+          <ComicCard
+            title={comic.title}
+            thumbnail={comic.thumbnail}
+            format={comic.format}
+            key={comic.id}
+            creators={comic.creators}
+            id={comic.id}
+          />
+        </div>
       );
     }
     return null;
